pyproject.toml
[project]
name = "pxc-demo"
version = "0.1.0"
dependencies = [
    "alembic",
    "fastapi[standard]",
    "jinja2",
    "jsonschema",
    "sqlmodel",
    "uvicorn[standard]",
    "wasmtime",
    # for markdown activity
    "markdown",
    "componentize-py",
    # LTI
    "PyJWT[crypto]",
    "jwcrypto",
]

[project.optional-dependencies]
dev = [
    "black",
    "datamodel-code-generator",
    "httpx",
    "mypy",
    "pylint",
    "pytest",
    "types-markdown",
    # LTI
    "responses",
]

[tool.setuptools.packages.find]
where = ["src"]

[tool.pylint.MAIN]
ignore-paths=[
    "src/pxc/notebook/frontend",
    "src/pxc/notebook/migrations/versions",
]

[tool.pylint.'MESSAGES CONTROL']
disable=[
    "duplicate-code",
    "fixme",
    "import-outside-toplevel",
    "invalid-name",
    "line-too-long",
    "missing-docstring",
    "protected-access",
    "redefined-outer-name",
    "too-few-public-methods",
    "too-many-arguments",
    "too-many-instance-attributes",
    "too-many-locals",
    "too-many-positional-arguments",
    "too-many-public-methods",
    "ungrouped-imports",
    "unused-argument",
    "wrong-import-order",
]

[tool.mypy]
plugins = ['pydantic.mypy']
warn_unused_configs = true
ignore_missing_imports = true
strict = true
exclude = [
    "src/pxc/notebook/frontend",
    "src/pxc/notebook/migrations/versions",
]

[tool.black]
extend-exclude = "src/pxc/notebook/frontend, src/pxc/notebook/migrations/versions"
